agent/requirements.txt
google-genai
google-cloud-firestore
mss
Pillow
pyautogui
pyaudio
python-dotenv
pygetwindow
pywinauto
websockets
pystray
pyperclip
